+++ rotkehlchen/chain/ethereum/modules/compound/v2/decoder.py
@@ -3,6 +3,10 @@ from typing import TYPE_CHECKING, Any
 
 from rotkehlchen.assets.asset import EvmToken
 from rotkehlchen.assets.utils import get_or_create_evm_token
+from rotkehlchen.chain.ethereum.modules.compound.constants import (
+    COMPTROLLER_PROXY_ADDRESS,
+    CPT_COMPOUND,
+)
 from rotkehlchen.chain.ethereum.utils import asset_normalized_value, token_normalized_value
 from rotkehlchen.chain.evm.decoding.constants import ERC20_OR_ERC721_TRANSFER
 from rotkehlchen.chain.evm.decoding.interfaces import DecoderInterface
@@ -23,7 +27,6 @@ from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.types import ChainID, ChecksumEvmAddress, EvmTokenKind, EvmTransaction
 from rotkehlchen.utils.misc import hex_or_bytes_to_address, hex_or_bytes_to_int
 
-from .constants import COMPTROLLER_PROXY_ADDRESS, CPT_COMPOUND
 from .utils import get_compound_underlying_token
 
 if TYPE_CHECKING: