+++ rotkehlchen/chain/evm/decoding/hop/constants.py
@@ -2,8 +2,7 @@ from typing import Final
 
 from rotkehlchen.chain.evm.decoding.types import CounterpartyDetails
 
-
-CPT_HOP: Final = 'hop-protocol'
+CPT_HOP: Final = 'hop'
 HOP_CPT_DETAILS: Final = CounterpartyDetails(
     identifier=CPT_HOP,
     label='Hop Protocol',