+++ rotkehlchen/db/upgrades/v42_v43.py
@@ -19,13 +19,24 @@ def _add_usd_price_nft_table(write_cursor: 'DBCursor') -> None:
     )
 
 
+@enter_exit_debug_log()
+def _change_hop_counterparty_value(write_cursor: 'DBCursor') -> None:
+    write_cursor.execute(
+        'UPDATE evm_events_info SET counterparty=? WHERE counterparty=?',
+        ('hop', 'hop-protocol'),
+    )
+
+
 @enter_exit_debug_log(name='UserDB v42->v43 upgrade')
 def upgrade_v42_to_v43(db: 'DBHandler', progress_handler: 'DBUpgradeProgressHandler') -> None:
     """Upgrades the DB from v42 to v43. This was in v1.34 release.
 
     - add usd_price to the nfts table
+    - change hop protocol counterparty value
     """
-    progress_handler.set_total_steps(1)
+    progress_handler.set_total_steps(2)
     with db.user_write() as write_cursor:
         _add_usd_price_nft_table(write_cursor)
+        progress_handler.new_step()
+        _change_hop_counterparty_value(write_cursor)
         progress_handler.new_step()